#157 stage 3 (phase 1): extract ort-free TableFormer core into tf_core
Groundwork for TableFormer in the browser: move the ONNX-free half of the
pipeline — the 448 encoder-input preprocessing, the autoregressive loop's
structure corrections (xcel→lcel, lcel-after-ucel→fcel) and bbox
bookkeeping (tag_H_buf / bboxes_to_merge), span merging, and the OTSL→grid
layout — into a new ocr-prep-gated tf_core module. The wasm build will run
these exact functions and delegate only the three ONNX graphs to ONNX
Runtime Web, the same one-implementation approach that keeps the browser
OCR/layout output matching native.

Native tableformer.rs (ml feature) now delegates to tf_core: preprocess
wraps tf_core::preprocess_input in an ort tensor; predict_otsl uses
tf_core::correct; predict_table_structure drives tf_core::BboxBook. Pure
code-motion — the owned-value KV-cache fast path and every ort call are
untouched, so native output is byte-identical. Unit tests cover the moved
pure functions (corrections, bbox bookkeeping, span build, argmax — noting
argmax ties to the last index via Rust max_by, kept verbatim).

Refs #157
